Api/requirements.txt
Flask
Flask-Cors
Jinja2
tensorflow
Keras
Pillow
gunicorn
certifi
